Reduce number of file transactions in configuration (#373)
* Update read_value_config to read from yml file

Before it was reading from a cached config object which goes stale if
the source file is updated.

* Update update_value_config to write back to file

Before updated data was just stored in memory

* Remove unused code

Remove unused imports and setup associated with old read and update

* Fix config manager unit tests

Split up the test into 3 tests
Catch the KeyErrors expected as KeyErrors are thrown for invalid keys
instead of returning None
Specify yaml loader as not doing so is deprecated

* Reformat with black

* Update test name

* Remove extraneous asserts

* Fix typo in raised Exception

* Replace call to get_value_config in ConfigurationManager

Remove get dependency on config_utils

* Replace update call in ConfigurationManager

* Remove unused imports

* Rename logging imports

* Add _create_if_not_exists and convert get and update into classmethods

Do not run bootstrap_environment if eva directory does not exist

* Remove unnecessary if clause

* Move udfs dir to constants

* Remove config empty string checks

When setting up config, we want to reset everything to prevent a corrupt
config

* Move config update to single file transaction

Before the read and write involved in an update were two separate file
transactions. This increases the likelihood of corruption if another
process modifies the file in between the read and write.

* Group variables and remove copytree argument

Copytree dirs_exist_ok does not exist in earlier versions of python

* Move upload_dir creation to right below eva_config_dir creation

* Reduce number of file transactions in bootstrap_env

Modify necessary configuration in single file transaction instead of 13

* Delete unused config_utils

* Add test for update_value

* Remove error handling for mode

If a yaml field does not exist we cannot do much so we assume all fields
exist. If the mode is something other than debug or release, default
logging level to debug.

* Use mode field in update test

Using datasets_dir field makes the other tests dump into two different
directories.