feat : Run EvaDB on Postgres (#922)
Summary : Made changes to provide an option that would run EVA on top of
Postgres server instead of SQLite. Created a CircleCI Job to validate
the changes. Currently the job is only run for Python 3.10, we can add
more configurations if needed.

Also fixed the UTs that were failing on CircleCI runs in default master.