fix tutorials 11-similarity-search-for-motif-mining.ipynb (#924)
Currently, the tutorials failed due to the error `database is locked`.
This is triggered by the `session.commit` in the
[catalog.upsert_udf_cost_catalog_entry](https://github.com/georgia-tech-db/evadb/blob/d9a10de0d1c1ea6084ee88d1517c61a54f63d530/evadb/executor/executor_utils.py#L50)
after the Yolo call. This error usually means that there is another
write to the sqlite
([reference](https://stackoverflow.com/questions/13895176/sqlalchemy-and-sqlite-database-is-locked)).
But we are not able to locate that now.

On the other hand, we are able to fix the problem by removing the
`batch_mem_size=12` at
[image_storage_engine.py](https://github.com/georgia-tech-db/evadb/blob/d9a10de0d1c1ea6084ee88d1517c61a54f63d530/evadb/storage/image_storage_engine.py#L30).
It is not clear why this fixes the `database is locked` issue. Need more
investigation to locate the root cause.

PS: it is also not clear where the current number 12 comes from.

---------

Co-authored-by: yulaicui <ycui96@gatech.edu>
Co-authored-by: Chitti Ankith <chittiankith@gmail.com>