feat: improve circle ci (#949)
Test folder reorg
```
test
|__ unit_tests
|__ integration_tests
     |__ short
     |__ long
```

For open PR, we will run unit tests and short integration tests. 

Once it is merged to staging branch, we will run long integration tests
across different versions and with both ray disabled or enabled.

Once it passes, it will be merged to master, which will run full test
(unit tests + all integration tests) with different python version and
with both ray disabled and enabled.

The testing for open PR is now around 8 mins, but majority of it is
installing package. I plan to also add cache for that but it is not
really working now.

---------

Co-authored-by: xzdandy <xzdandy@gmail.com>