Add a model forecasting notebook in tutorials (#1118)
Add a notebook for forecasting.

Blocked by
- [x] #1117 Bypassed by installing evadb from a commit before the PR.
- [x] #1119 Bypassed by using a nested SQL query, will improve later.

This PR adds a basic version of notebook for forecasting. We have also
found several issues we need to improve in later PRs:
**EvaDB related:**
1. Nested SQL query must have AS, which is redundant.
2. Can not order by a column which is the function's output from the
projection's target list.
3. When we are tuning the parameters that used for training, we need to
run the drop function every time.

**Statsforecast integration related:**
1. Math domain error with statsforecast, when we want to forecast
property prices for different postcode.
2. Auto inferring frequency never works.
3. Changing the frequency does not seem to affect the output of the
forecasting value, which is unexpected.
5. The catalog column can only be one column, so we need to train three
different models for postcode, bedrooms, and propertytype. We should be
able to train one model for every combination of these columns.