feat: integrate with pgvector (#1153)
I will need some feedback on the design here. @xzdandy @gaurav274 

This PR reflects my initial design for in-database index features like
pgvector. My initial idea is to offload as much as possible to the
native database. When we create an index and do the index scan, we
simply push an emulated query to the underlying database. While this is
doable, this will introduce separate implementation paths in different
components including the optimizer and the executor.
 
Other than the current design, another option is to reuse the
third-party vector integration interface. There are also a few details
that I am not clear about
* For other vector libraries, we keep an index entry in our own catalog.
Do we still maintain that even for pgvector? In the original
implementation, the index catalog entry is linked to a table catalog
entry. Because in this case, the table is inside Postgres, we also need
to change the implementation here a little bit.
* ~~(Issue related to performance). Following the current create index
implementation, data will be fetched from Postgres first but it is not
really needed. The create index will anyway run inside Postgres.~~
* ~~The current vector index scan is implemented based on `_row_id`.
When it is scanning data from Postgres, we need to figure out a way to
populate `_row_id` for the native database engine.~~